requirements.txt
fastapi
uvicorn
pandas
PyPDF2
pymongo
transformers
pymilvus
langchain
dashscope
python-docx
rapidocr-onnxruntime
unstructured
pymupdf
tqdm
werkzeug
langchain-community
seaborn
zhipuai
gradio
pillow
selenium
pyecharts
snapshot-selenium
openpyxl